solarmind/requirements.txt
# SolarMind AI — Python Dependencies
# Core
python-dotenv>=1.0,<2.0
structlog>=24.0,<25.0
pathlib2>=2.3; python_version < "3.4"

# Data
pandas>=2.1,<3.0
numpy>=1.26,<2.0
pyarrow>=15.0,<16.0
openpyxl>=3.1,<4.0

# ML
xgboost>=2.0,<3.0
scikit-learn>=1.4,<2.0
shap>=0.44,<1.0
lightgbm>=4.3.0
catboost>=1.2.0
imbalanced-learn>=0.12.0

# GenAI / LLM
openai>=1.10,<2.0
pydantic>=2.5,<3.0

# RAG
chromadb>=0.4,<1.0

# Agent
langgraph>=0.0.40,<1.0
langchain-core>=0.1,<1.0

# API
fastapi>=0.110,<1.0
uvicorn>=0.27,<1.0
python-jose[cryptography]>=3.3,<4.0
prometheus-client>=0.20,<1.0
websockets>=12.0,<13.0
httpx>=0.27,<1.0

# Dashboard (installed via npm separately)
# react, recharts, tailwindcss

# Testing
pytest>=8.0,<9.0
pytest-asyncio>=0.23,<1.0
